requirements.txt
opencv-python~=4.9.0.80
open-clip-torch~=2.24.0
transformers~=4.37.2
